+++ src/main/java/net/jcm/vsch/util/ShipUtils.java
@@ -1,28 +1,16 @@
 package net.jcm.vsch.util;
 
-import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.level.block.Rotation;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.phys.Vec3;
 import org.joml.Quaterniond;
 import org.joml.Vector3d;
-import org.joml.Vector3i;
-import org.valkyrienskies.core.api.ships.ServerShip;
 import org.valkyrienskies.core.api.ships.Ship;
 import org.valkyrienskies.core.api.ships.properties.ShipTransform;
 import org.valkyrienskies.core.apigame.physics.PhysicsEntityServer;
 import org.valkyrienskies.core.apigame.world.ServerShipWorldCore;
 import org.valkyrienskies.core.impl.game.ships.ShipObjectServerWorld;
 import org.valkyrienskies.core.impl.game.ships.ShipTransformImpl;
-import org.valkyrienskies.core.util.datastructures.DenseBlockPosSet;
-import org.valkyrienskies.mod.common.VSGameUtilsKt;
-import org.valkyrienskies.mod.common.assembly.ShipAssemblyKt;
 import org.valkyrienskies.mod.common.util.EntityDraggingInformation;
 import org.valkyrienskies.mod.common.util.IEntityDraggingInformationProvider;
-import org.valkyrienskies.mod.common.util.VectorConversionsMCKt;
-import org.valkyrienskies.mod.util.RelocationUtilKt;
 
 import java.util.Objects;
 
@@ -55,99 +43,5 @@ public class ShipUtils {
         EntityDraggingInformation information = provider.getDraggingInformation();
         if (!information.isEntityBeingDraggedByAShip()) return false;
         return Objects.equals(information.getLastShipStoodOn(), id);
-    }
-
-    /**
-     * <b>Warning:</b> This function is still affected by the height bug that other addons have,
-     * and so can't be used in space.
-     *
-     * <p></p>
-     *
-     * <code>dx</code>, <code>dy</code> and <code>dz</code>
-     * Are the distance from center block to be assembled. Added in both directions, so
-     * <code>(center.x - dx)</code> ... <code>(center.x + dx)</code>.
-     * 
-     * @see #assembleBlock(ServerLevel, BlockPos)
-     * 
-     * @return The ship that was created, or null if it wasn't created
-     */
-    public static ServerShip assembleBlocks(ServerLevel level, BlockPos center, int dx, int dy, int dz) {
-        if (dx < 0 || dy < 0 || dz < 0) return null;
-
-        DenseBlockPosSet set = new DenseBlockPosSet();
-        for (int x = -dx; x <= dx; x++) {
-            for (int y = -dy; y <= dy; y++) {
-                for (int z = -dz; z <= dz; z++) {
-                    BlockPos newBlock = new BlockPos(center.getX() + x, center.getY() + y, center.getZ() + z);
-                    if (!level.getBlockState(newBlock).isAir()) {
-                        set.add(VectorConversionsMCKt.toJOML(newBlock));
-                    }
-                }
-            }
-        }
-
-        // Otherwise we get massless ships
-        if (set.isEmpty()) return null;
-
-        double e = (double) new java.util.function.Supplier<Number>() {
-            public Number get() {
-                Ship ship = null;
-                if (ship instanceof ServerShip) {
-                    return ((ServerShip) ship).getTransform().getShipToWorldScaling().x();
-                }
-                return 0;
-            }
-        }.get().doubleValue();
-
-
-        Vec3 test = (
-                new java.util.function.Supplier<Vec3>() {
-                    public Vec3 get() {
-                        Ship ship = null;
-                        if (ship != null) {
-                            if (ship.getWorldAABB() != null) {
-                                return VectorConversionsMCKt.toMinecraft(ship.getWorldAABB().center(new Vector3d(0, 0, 0)));
-                            }
-                        }
-                        return new Vec3(0, 0, 0);
-                    }
-                }.get()
-        );
-
-        return ShipAssemblyKt.createNewShipWithBlocks(center, set, level);
-    }
-
-    /**
-     * This function IS safe to use in space, but only assembles a single block.
-     *
-     * @see #assembleBlocks(ServerLevel, BlockPos, int, int, int)
-     *
-     * @return The ship that was created, or null if it wasn't created
-     */
-    public static ServerShip assembleBlock(ServerLevel level, BlockPos blockPos) {
-        BlockState blockState = level.getBlockState(blockPos);
-
-        if (VSGameUtilsKt.getShipManagingPos(level, blockPos) != null) return null;
-
-        if (blockState.isAir()) return null;
-
-        String dimensionId = VSGameUtilsKt.getDimensionId(level);
-
-
-        ServerShip serverShip =
-                VSGameUtilsKt.getShipObjectWorld(level)
-                        .createNewShipAtBlock(
-                                VectorConversionsMCKt.toJOML(blockPos),
-                                false,
-                                1.0,
-                                dimensionId
-                        );
-
-        BlockPos centerPos = VectorConversionsMCKt.toBlockPos(serverShip.getChunkClaim().getCenterBlockCoordinates(VSGameUtilsKt.getYRange(level), new Vector3i()));
-
-        // Move the block from the world to a ship
-        RelocationUtilKt.relocateBlock(level, blockPos, centerPos, true, serverShip, Rotation.NONE);
-
-        return serverShip;
     }
 }
